projeto_folha_de_ponto/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
python-dateutil==2.8.2
pdfkit==1.0.0
Jinja2==3.1.2